+++ exporter/postgres_exporter_integration_test.go
@@ -21,7 +21,6 @@ package exporter
 import (
 	"fmt"
 	"os"
-	"strings"
 	"testing"
 
 	_ "github.com/lib/pq"
@@ -43,7 +42,7 @@ func (s *IntegrationSuite) SetUpSuite(c *C) {
 	dsn := os.Getenv("DATA_SOURCE_NAME")
 	c.Assert(dsn, Not(Equals), "")
 
-	exporter := NewExporter(strings.Split(dsn, ","), promslog.NewNopLogger())
+	exporter := NewExporter(dsn, promslog.NewNopLogger())
 	c.Assert(exporter, NotNil)
 	// Assign the exporter to the suite
 	s.e = exporter
@@ -60,26 +59,24 @@ func (s *IntegrationSuite) TestAllNamespacesReturnResults(c *C) {
 		}
 	}()
 
-	for _, dsn := range s.e.dsn {
-		// Open a database connection
-		server, err := NewServer(dsn)
-		c.Assert(server, NotNil)
-		c.Assert(err, IsNil)
+	// Open a database connection
+	server, err := NewServer(s.e.dsn)
+	c.Assert(server, NotNil)
+	c.Assert(err, IsNil)
 
-		// Do a version update
-		err = s.e.checkMapVersions(ch, server)
-		c.Assert(err, IsNil)
+	// Do a version update
+	err = s.e.checkMapVersions(ch, server)
+	c.Assert(err, IsNil)
 
-		// This should never happen in our test cases.
-		errMap := queryNamespaceMappings(ch, server)
-		if !c.Check(len(errMap), Equals, 0) {
-			fmt.Println("## NAMESPACE ERRORS FOUND")
-			for namespace, err := range errMap {
-				fmt.Println(namespace, ":", err)
-			}
+	// This should never happen in our test cases.
+	errMap := queryNamespaceMappings(ch, server)
+	if !c.Check(len(errMap), Equals, 0) {
+		fmt.Println("## NAMESPACE ERRORS FOUND")
+		for namespace, err := range errMap {
+			fmt.Println(namespace, ":", err)
 		}
-		server.Close()
 	}
+	server.Close()
 }
 
 // TestInvalidDsnDoesntCrash tests that specifying an invalid DSN doesn't crash
@@ -94,12 +91,12 @@ func (s *IntegrationSuite) TestInvalidDsnDoesntCrash(c *C) {
 	}()
 
 	// Send a bad DSN
-	exporter := NewExporter([]string{"invalid dsn"}, promslog.NewNopLogger())
+	exporter := NewExporter("invalid dsn", promslog.NewNopLogger())
 	c.Assert(exporter, NotNil)
 	exporter.scrape(ch)
 
 	// Send a DSN to a non-listening port.
-	exporter = NewExporter([]string{"postgresql://nothing:nothing@127.0.0.1:1/nothing"}, promslog.NewNopLogger())
+	exporter = NewExporter("postgresql://nothing:nothing@127.0.0.1:1/nothing", promslog.NewNopLogger())
 	c.Assert(exporter, NotNil)
 	exporter.scrape(ch)
 }
@@ -117,7 +114,7 @@ func (s *IntegrationSuite) TestUnknownMetricParsingDoesntCrash(c *C) {
 	dsn := os.Getenv("DATA_SOURCE_NAME")
 	c.Assert(dsn, Not(Equals), "")
 
-	exporter := NewExporter(strings.Split(dsn, ","), promslog.NewNopLogger())
+	exporter := NewExporter(dsn, promslog.NewNopLogger())
 	c.Assert(exporter, NotNil)
 
 	// Convert the default maps into a list of empty maps.
@@ -149,7 +146,7 @@ func (s *IntegrationSuite) TestExtendQueriesDoesntCrash(c *C) {
 	c.Assert(dsn, Not(Equals), "")
 
 	exporter := NewExporter(
-		strings.Split(dsn, ","),
+		dsn,
 		promslog.NewNopLogger(),
 		WithUserQueriesPath("../user_queries_test.yaml"),
 	)
@@ -163,7 +160,7 @@ func (s *IntegrationSuite) TestAutoDiscoverDatabases(c *C) {
 	dsn := os.Getenv("DATA_SOURCE_NAME")
 
 	exporter := NewExporter(
-		strings.Split(dsn, ","),
+		dsn,
 		promslog.NewNopLogger(),
 	)
 	c.Assert(exporter, NotNil)